+++ src/ui/hooks/docs.ts
@@ -1,6 +1,5 @@
 import recommendMd from '../assets/vendors-recommand.md?raw';
 import readMeMd from '../../../README.md?raw';
-import upgradeMd from '../../../UPGRADE.md?raw';
 import { useEffect, useState } from 'react';
 import { api } from '../api/client';
 
@@ -52,4 +51,4 @@ function createDocHookFunc(markdown: string, apiFunc: Function) {
 
 export const useRecomandVendors = createDocHookFunc(recommendMd, api.getRecommendVendorsMarkdown);
 export const useReadme = createDocHookFunc(readMeMd, api.getReadmeMarkdown);
-export const useUpgradeNotes = createDocHookFunc(upgradeMd, api.getUpgradeMarkdown);
+export const useUpgradeNotes = createDocHookFunc('', api.getUpgradeMarkdown);